:truck: Rename MyTravelAuthorizationEditLayout-DetailsTab to MyTravelAuthorizationsEditDetailsPage.
Why? I'm attempting to break out "pages" as think you can route directly to, and things that wrap those as "layouts"
