+++ web/src/modules/travel-authorizations/components/my-travel-authorizations-edit-details-page/details-form-card/MultiDestinationStopsSection.vue
