+++ web/src/modules/travel-authorizations/layouts/MyTravelAuthorizationEditLayout.vue
@@ -17,7 +17,7 @@
     </template>
 
     <v-tabs v-model="tab">
-      <v-tab :to="{ name: 'MyTravelAuthorizationEditLayout-DetailsTab', params: { formId } }"
+      <v-tab :to="{ name: 'MyTravelAuthorizationsEditDetailsPage', params: { formId } }"
         >Details</v-tab
       >
       <v-tab
@@ -70,7 +70,7 @@ export default {
     // Hacky thing to refresh travel authorization after user edits the estimates in the Estimate tab.
     // This does a wizard of oz style, silent, background refresh.
     $route(to) {
-      if (to.name === "MyTravelAuthorizationEditLayout-DetailsTab") {
+      if (to.name === "MyTravelAuthorizationsEditDetailsPage") {
         this.loadCurrentTravelAuthorizationSilently(this.formId)
       }
     },
