+++ web/src/modules/travel-authorizations/layouts/README.md
@@ -1,0 +1,3 @@
+# Travel Authorization Layouts
+
+Layouts are components that wrap a page or collection of pages, then can't be routed to directly.
