+++ web/src/modules/travel-authorizations/pages/MyTravelAuthorizationsEditDetailsPage.vue
@@ -32,14 +32,14 @@
 </template>
 
 <script>
-import PurposeFormCard from "./details-tab/PurposeFormCard"
-import DetailsFormCard from "./details-tab/DetailsFormCard"
-import ApprovalsFormCard from "./details-tab/ApprovalsFormCard"
+import PurposeFormCard from "@/modules/travel-authorizations/components/my-travel-authorizations-edit-details-page/PurposeFormCard"
+import DetailsFormCard from "@/modules/travel-authorizations/components/my-travel-authorizations-edit-details-page/DetailsFormCard"
+import ApprovalsFormCard from "@/modules/travel-authorizations/components/my-travel-authorizations-edit-details-page/ApprovalsFormCard"
 
-import SaveDraftButton from "./details-tab/SaveDraftButton"
+import SaveDraftButton from "@/modules/travel-authorizations/components/my-travel-authorizations-edit-details-page/SaveDraftButton"
 
 export default {
-  name: "DetailsTab",
+  name: "MyTravelAuthorizationsEditDetailsPage",
   components: {
     PurposeFormCard,
     DetailsFormCard,
