+++ web/src/modules/travel-authorizations/pages/README.md
@@ -1,0 +1,9 @@
+# Travel Authorization Pages
+
+Pages are components that can be routed to directly.
+Component names should end in `Page`.
+Page specific components should be in the components/name-of-page-component directory.
+
+e.g.
+- `web/src/modules/travel-authorizations/pages/MyTravelAuthorizationsEditDetailsPage.vue`
+has specific components in `web/src/modules/travel-authorizations/components/my-travel-authorizations-edit-details-page`
